requirements.txt
fastapi>=0.109.0,<0.110.0
uvicorn>=0.27.0,<0.28.0
pydantic>=2.6.0,<3.0.0
httpx>=0.26.0,<0.28.0
dnspython>=2.6.1,<2.7.0
ulid-py>=1.1.0,<1.2.0
packaging>=23.2,<25.0